+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Import the contents of CSV-BUILDINGS-MULTIPLE into a postgres table that can be joined on the footprints
 
+### Changed
+- `batch3dfy` console script -> `batch3dfier`
+
 ## [0.6.0] - 2017-08-08
 ### Added
 - support for python 3.4-3.6