+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Import the contents of CSV-BUILDINGS-MULTIPLE into a postgres table that can be joined on the footprints
+- Log file (just basic info)
+- Wheel
 
 ### Changed
 - `batch3dfy` console script -> `batch3dfier`